使用 valgrind 3.14版本，make 能够编译通过

+++ valgrind/coregrind/m_gdbserver/remote-utils.c
@@ -1158,7 +1158,7 @@ void prepare_resume_reply (char *buf, char status, unsigned char sig)
          VG_(sprintf) (buf, "%s:%x;",
                        valgrind_stopped_before_syscall ()
                        ? "syscall_entry" : "syscall_return",
-                       (UInt)valgrind_stopped_by_syscall ());
+                       valgrind_stopped_by_syscall ());
          buf += strlen (buf);
       }
